backend/requirements.txt
Flask
supabase
flask-cors
python-dotenv